# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics Inventory API — multi-tenant inventory SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      parameters:
        - { in: query, name: status, required: false, schema: { type: string } }
        - { in: query, name: supplierId, required: false, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PurchaseOrder" }
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePurchaseOrderPayload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrderDetail" }
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrderDetail" }
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePurchaseOrderPayload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrderDetail" }
    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
  /purchase-orders/{id}/pdf:
    get:
      operationId: downloadPurchaseOrderPdf
      tags: [purchase-orders]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Purchase order PDF binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /purchase-orders/{id}/status:
    patch:
      operationId: updatePurchaseOrderStatus
      tags: [purchase-orders]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrderDetail" }
  /purchase-orders/{id}/return:
    post:
      operationId: returnPurchaseOrder
      tags: [purchase-orders]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReturnOrderPayload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrderDetail" }
...
components:
  schemas:
...
    OrderLineInput:
      type: object
      properties:
        itemId: { type: integer }
        quantity: { type: number }
        unitPrice: { type: number }
        taxRate: { type: number }
        description: { type: ["string", "null"] }
      required: [itemId, quantity, unitPrice, taxRate]
    OrderLine:
      type: object
      properties:
        id: { type: integer }
        itemId: { type: integer }
        itemName: { type: string }
        sku: { type: string }
        variantOptions:
          type: ["object", "null"]
          additionalProperties: { type: string }
        quantity: { type: number }
        quantityShipped: { type: number }
        quantityReceived: { type: number }
        unitPrice: { type: number }
        taxRate: { type: number }
        lineSubtotal: { type: number }
        lineTax: { type: number }
        lineTotal: { type: number }
        description: { type: ["string", "null"] }
        trackBatches:
          type: boolean
          description: True when the line item has batch tracking enabled. Receipts on tracked lines must capture batches; shipments on tracked lines must pick from existing batches.
      required: [id, itemId, itemName, sku, variantOptions, quantity, quantityShipped, quantityReceived, unitPrice, taxRate, lineSubtotal, lineTax, lineTotal, description, trackBatches]
...
    PurchaseOrder:
      type: object
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        supplierId: { type: integer }
        supplierName: { type: string }
        warehouseId: { type: integer }
        warehouseName: { type: string }
        status: { type: string }
        orderDate: { type: string }
        expectedDeliveryDate: { type: ["string", "null"] }
        subtotal: { type: number }
        taxTotal: { type: number }
        total: { type: number }
        amountPaid: { type: number }
        balanceDue: { type: number }
        notes: { type: ["string", "null"] }
        jobWorkReceiptId: { type: ["integer", "null"] }
        jobWorkOrderId: { type: ["integer", "null"] }
        jwoNumber: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, orderNumber, supplierId, supplierName, warehouseId, warehouseName, status, orderDate, expectedDeliveryDate, subtotal, taxTotal, total, amountPaid, balanceDue, notes, jobWorkReceiptId, jobWorkOrderId, jwoNumber, createdAt]
...
    PurchaseOrderDetail:
      type: object
      properties:
        order: { $ref: "#/components/schemas/PurchaseOrder" }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OrderLine" }
        goodsReceipts:
          type: array
          items: { $ref: "#/components/schemas/GoodsReceipt" }
      required: [order, lines, goodsReceipts]
    GoodsReceiptLine:
      type: object
      properties:
        id: { type: integer }
        goodsReceiptId: { type: integer }
        purchaseOrderLineId: { type: integer }
        itemName: { type: string }
        sku: { type: string }
        quantity: { type: number }
      required: [id, goodsReceiptId, purchaseOrderLineId, itemName, sku, quantity]
    GoodsReceipt:
      type: object
      properties:
        id: { type: integer }
        purchaseOrderId: { type: integer }
        receiptNumber: { type: string }
        receivedDate: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/GoodsReceiptLine" }
      required: [id, purchaseOrderId, receiptNumber, receivedDate, status, notes, createdAt, lines]
...
    CreatePurchaseOrderPayload:
      type: object
      properties:
        supplierId: { type: integer }
        warehouseId: { type: integer }
        orderDate: { type: string }
        expectedDeliveryDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OrderLineInput" }
      required: [supplierId, warehouseId, orderDate, lines]
    UpdateOrderStatusBody:
      type: object
      properties:
        status: { type: string }
      required: [status]
    ReturnOrderPayload:
      type: object
      properties:
        notes: { type: ["string", "null"] }
...
    UpdatePurchaseOrderPayload:
      type: object
      properties:
        supplierId: { type: integer }
        warehouseId: { type: integer }
        orderDate: { type: string }
        expectedDeliveryDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OrderLineInput" }
...